requirements.txt
paho-mqtt
Adafruit_DHT